- Reuse route.RoutesResult via sync.Pool
- Improve util.Cache


Before

```sh
% go test -bench=Route -benchmem -memprofile=mem.prof -cpuprofile=cpu.prof ./pkg/route/...
goos: darwin
goarch: arm64
pkg: github.com/fasthttpd/fasthttpd/pkg/route
BenchmarkRoutes_Equal-8          	23975185	        45.16 ns/op	     128 B/op	       1 allocs/op
BenchmarkCachedRoutes_Equal-8    	12914530	        99.09 ns/op	       5 B/op	       1 allocs/op
BenchmarkRoutes_Prefix-8         	26712596	        45.34 ns/op	     128 B/op	       1 allocs/op
BenchmarkCachedRoutes_Prefix-8   	12729596	        94.67 ns/op	      24 B/op	       1 allocs/op
BenchmarkRoutes_Regexp-8         	12622537	        96.67 ns/op	     132 B/op	       1 allocs/op
BenchmarkCachedRoutes_Regexp-8   	12501996	        95.00 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/fasthttpd/fasthttpd/pkg/route	8.056s
```

After

```sh
% go test -bench=Route -benchmem -memprofile=mem.prof -cpuprofile=cpu.prof ./pkg/route/...
goos: darwin
goarch: arm64
pkg: github.com/fasthttpd/fasthttpd/pkg/route
BenchmarkRoutes_Equal-8          	327433689	        3.577 ns/op	       0 B/op	       0 allocs/op
BenchmarkCachedRoutes_Equal-8    	29450773	        40.12 ns/op	       1 B/op	       1 allocs/op
BenchmarkRoutes_Prefix-8         	281970042	        4.253 ns/op	       0 B/op	       0 allocs/op
BenchmarkCachedRoutes_Prefix-8   	28136815	        39.80 ns/op	       1 B/op	       1 allocs/op
BenchmarkRoutes_Regexp-8         	20191369	        55.48 ns/op	       0 B/op	       0 allocs/op
BenchmarkCachedRoutes_Regexp-8   	28502349	        41.00 ns/op	       1 B/op	       1 allocs/op
PASS
ok  	github.com/fasthttpd/fasthttpd/pkg/route	8.344s
```